requirements.txt
PyQt5-Qt5>=5.15.2
PyQt5-sip>=12.8.1
PyMuPDF==1.24.12
requests==2.31.0
packaging==23.2
pyinstaller==6.3.0